+++ resources/views/bids/edit-bid.blade.php
@@ -65,9 +65,16 @@
                             'contractors'=>$contractors
                         ])
 
-                        @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot().'::bids/shared/file-list',[
-                           'bid'=>$bid,'b_edit'=>true
-                        ])
+                        <div class="mt-5">
+                            @include('shared.documents.project_documents',
+                                  ['project_files'=>$bid->bid_files,'file_type' => 'bid_files',
+                                    'image_gallery_url' => route('scm-bid.bid.image_gallery',['single_bid'=>$bid])
+                                  ])
+
+                            @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot() . '::bids/shared/bid-file-context-scripts',
+                                       ['b_edit' => true])
+                        </div>
+
 
                         <div class="mt-5">
                             @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot().'::bids/edit/bid-add-files',[