+++ resources/views/bids/show-bid.blade.php
@@ -1,6 +1,7 @@
 @php
     /**
      * @var \Scm\PluginBid\Models\ScmPluginBidSingle $bid
+     * @var bool $b_edit
      */
 @endphp
 
@@ -114,16 +115,13 @@
                             </div>
                         </div>
 
-
-
-                        @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot().'::bids/shared/file-list',[
-                           'bid'=>$bid,'b_edit'=>false
-                        ])
-
                         <div class="row mt-4">
-                            @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot().'::bids/show/file-images',[
-                                  'bid'=>$bid
-                               ])
+                            @include('shared.documents.project_documents',
+                                   ['project_files'=>$bid->bid_files,'file_type' => 'bid_files',
+                                     'image_gallery_url' => route('scm-bid.bid.image_gallery',['single_bid'=>$bid])
+                                   ])
+                            @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot() . '::bids/shared/bid-file-context-scripts',
+                                        ['b_edit' => $b_edit])
                         </div>
 
                     </div> <!-- /card-body -->